requirements.txt
absl-py==2.1.0
afinn==0.1
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.7.0
blis==0.7.11
boto3==1.34.89
botocore
build==1.1.1
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.1
confection==0.1.4
contourpy==1.2.0
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.4
datasets==2.16.1
decorator==5.1.1
deepdiff==6.7.1
Deprecated==1.2.14
dill==0.3.7
distro==1.9.0
dnspython==2.5.0
docker-pycreds==0.4.0
emoji==2.10.1
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl#sha256=ab70aeb6172cde82508f7739f35ebc9918a3d07debeed637403c8f794ba3d3dc
entrypoints==0.4
et-xmlfile==1.1.0
evaluate==0.4.1
exceptiongroup==1.2.0
executing==2.0.1
Faker==23.2.0
fastapi==0.110.0
favicon==0.7.0
filelock==3.13.1
filetype==1.2.0
FlashRank==0.2.0
Flask==3.0.2
flatbuffers==24.3.7
fonttools==4.49.0
frozendict==2.4.0
frozenlist==1.4.1
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.41
google-ai-generativelanguage==0.4.0
google-api-core==2.18.0
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-generativeai==0.4.1
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.62.1
grpcio-status==1.62.1
gunicorn==21.2.0
h11==0.14.0
h5py==3.10.0
htbuilder==0.6.2
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.20.3
humanfriendly==10.0
humanize==4.9.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
ipython==8.18.1
ipywidgets==8.1.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
kubernetes
langchain==0.1.1
langchain-community==0.0.14
langchain-core==0.1.14
langchain-openai==0.0.3
langcodes==3.3.0
langdetect==1.0.9
langkit==0.0.30
langsmith==0.0.83
litellm==1.34.37
lxml==5.1.0
Mako==1.3.2
Markdown==3.5.2
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.4
marshmallow==3.20.2
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
merkle-json==1.0.0
millify==0.1.1
mmh3==4.1.0
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
munch==4.0.0
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
oauthlib==3.2.2
onnxruntime==1.17.1
openai==1.9.0
openpyxl==3.1.2
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
ordered-set==4.1.0
orjson==3.9.15
overrides==7.7.0
packaging==23.2
pandas==2.2.0
parso==0.8.3
pdf2image==1.17.0
pdfminer==20191125
phonenumbers==8.13.30
pika==0.12.0
pillow==10.2.0
platformdirs==3.11.0
posthog==3.5.0
preshed==3.0.9
presidio-analyzer==2.2.353
presidio-anonymizer==2.2.354
prometheus_client==0.20.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.23.4
psutil==5.9.8
pulsar-client==3.4.0
pure-eval==0.2.2
pyarrow==15.0.0
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycryptodome==3.20.0
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments==2.17.2
pymdown-extensions==10.7
pymongo==4.6.1
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
pyparsing==3.1.1
pypdf==4.1.0
pyphen==0.14.0
PyPika==0.48.9
pyproject_hooks==1.0.0
pyreadline3==3.4.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
python-iso639==2024.2.7
python-magic==0.4.27
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.6.2
referencing==0.32.1
regex==2023.12.25
requests
requests-file==2.0.0
requests-oauthlib==1.3.1
responses
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
s3transfer==0.10.1
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.12.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk
sentry-sdk
seqeval==1.2.2
setproctitle==1.3.3
simpletransformers==0.63.11
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.25
srsly==2.4.8
st-annotated-text==4.0.1
stack-data==0.6.3
starlette==0.36.3
streamlit==1.30.0
streamlit-aggrid==0.3.4.post3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.0
streamlit-embedcode==0.1.2
streamlit-extras==0.4.0
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.3
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
textstat==0.7.3
thinc==8.2.3
threadpoolctl==3.2.0
tiktoken==0.5.2
tldextract==5.1.1
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.39.3
trio==0.22.2
trio-websocket==0.10.3
trulens==0.13.4
trulens-eval==0.22.2
typer==0.9.0
types-requests
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
unstructured==0.12.5
unstructured-client==0.21.1
urllib3
uvicorn==0.27.1
vaderSentiment==3.3.2
validators==0.22.0
wandb==0.16.2
wasabi==1.1.2
watchdog==3.0.0
watchfiles==0.21.0
wcwidth==0.2.13
weasel==0.3.4
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
whylabs-client==0.5.10
whylogs==1.3.23
whylogs-sketching==3.4.1.dev3
widgetsnbextension==4.0.10
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
